Config unix scripts

+++ config/unix/fedora/install.sh
@@ -2,9 +2,9 @@
 
 for arg in "$@"
 do
-  if [ "$arg" == "--help" ] || [ "$arg" == "-h" ]
-  then
-    cat << EOF
+    if [ "$arg" == "--help" ] || [ "$arg" == "-h" ]
+    then
+cat << EOF
 
 Fedora Install Script
 
@@ -26,33 +26,26 @@ EOF
   if [ "$arg" == "--dev" ] || [ "$arg" == "-d" ]
   then
     DEV=1
-  else
-    DEV=0
   fi
   if [ "$arg" == "--clean" ] || [ "$arg" == "-c" ]
   then
     CLEAN=1
-  else
-    CLEAN=0
   fi
   if [ "$arg" == "--installmissing" ] || [ "$arg" == "-im" ]
   then
     INSTALLMISSINGPACKAGES=1
-  else
-    INSTALLMISSINGPACKAGES=0
   fi
 done
 
-if [ $MEDIA_SUBSHELL_ACTIVE == 1 ]
+if [ $MEDIA_SUBSHELL_ACTIVE ] && [ "$MEDIA_SUBSHELL_ACTIVE" -eq "1" ]
 then
   deactivate
-  export MEDIA_SUBSHELL_ACTIVE=0
+  export MEDIA_SUBSHELL_ACTIVE="0"
 fi
 
 FEDORADIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
-echo $FEDORADIR
 CONFIGDIR=$FEDORADIR
-while [ $(basename $CONFIGDIR) != "/" ] && [ $(basename $CONFIGDIR) != "config" ]
+while [ "$(basename $CONFIGDIR)" != "/" ] && [ "$(basename $CONFIGDIR)" != "config" ]
 do
   CONFIGDIR="$(dirname "$CONFIGDIR")"
 done
@@ -61,10 +54,11 @@ source $CONFIGDIR/unix/env.sh
 original_dir=$(pwd)
 cd $MEDIA_PROJECT_DIR
 
-if [ $CLEAN == 1 ]
+if [ $CLEAN ] && [ $CLEAN -eq 1 ]
   then
   rm -rf $MEDIA_PROJECT_DIR/.venv
 fi
+
 
 function isinstalled {
   if [ "$(pip list | grep "$@" 2>/dev/null)" ] || yum list installed "$@" >/dev/null 2>&1; then
@@ -81,30 +75,31 @@ for dependency in ${dependencies} ; do
   if isinstalled $dependency
   then
     echo "requirement satisfied:" $dependency
-  elif [ $INSTALLMISSINGPACKAGES == 1 ]
+
+  elif [ $INSTALLMISSINGPACKAGES -eq 1 ]
   then
     echo "installing " $dependency
     PIPFAILED=0
     YUMFAILED=0
-    if [ $(id -u) -ne 0 ]
-    then
-      if pip install --user $dependency >/dev/null 2>&1
+    if [ "$(id -u)" -ne "0" ]
       then
-        echo "pip install succeeded"
-        continue
+        if pip install --user $dependency >/dev/null 2>&1
+        then
+          echo "pip install succeeded"
+          continue
+        else
+          echo "pip install failed"
+          PIPFAILED=1
+        fi
       else
-        echo "pip install failed"
-        PIPFAILED=1
-      fi
-    else
-      if sudo -u $USER pip install --user $dependency >/dev/null 2>&1
-      then
-        echo "pip install succeeded"
-        continue
-      else
-        echo "pip install failed"
-        PIPFAILED=1
-      fi
+        if sudo -u $USER pip install --user $dependency >/dev/null 2>&1
+        then
+          echo "pip install succeeded"
+          continue
+        else
+          echo "pip install failed"
+          PIPFAILED=1
+        fi
     fi
     if yum install $dependency >/dev/null 2>&1
     then
@@ -114,11 +109,12 @@ for dependency in ${dependencies} ; do
       echo "yum install failed"
       YUMFAILED=1
     fi
-    if [ $PIPFAILED == 1 ] && [ $YUMFAILED == 1 ]
+    if [ $PIPFAILED -eq 1 ] && [ $YUMFAILED -eq 1 ]
       then
         FAILED=1
         continue
     fi
+
   else
     echo $dependency "not installed. Please install as root."
     FAILED=1
@@ -126,7 +122,7 @@ for dependency in ${dependencies} ; do
   fi
 done
 
-if [ $FAILED == 1 ]
+if [ $FAILED ] [ $FAILED -eq 1 ]
   then
     echo "Build failed. Check your dependencies, and install with sudo if you can."
     return 1
@@ -140,21 +136,21 @@ source $CONFIGDIR/unix/env.sh
 export PIPENV_VENV_IN_PROJECT=true
 
 # Install dev tools if dev is true
-if [ $DEV == 1 ]
+if [ $DEV ] && [ $DEV -eq 1 ]
 then
     pipenv install --dev
 else
     pipenv install
 fi
 
-if [ $INSTALLMISSINGPACKAGES == 1 ]
+if [ $INSTALLMISSINGPACKAGES ] && [ $INSTALLMISSINGPACKAGES -eq 1 ]
 then
   source $FEDORADIR/install_pyside.sh --installmissing
 else
   source $FEDORADIR/install_pyside.sh
 fi
 
-if [ "$(id -u)" == "0" ]
+if [ $(id -u) -eq 0 ]
 then
   echo "Install was run as root. Changing permissions in directory:" $MEDIA_PROJECT_DIR/.venv
   sudo chmod 777 .
